♻️ 関数を移動
